removing tkinter app
