+++ CLAUDE.md
@@ -7,7 +7,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 - [docs/profile.md](docs/profile.md) — Profile parameters, files, tolerance, handle connectors
 - [docs/system.md](docs/system.md) — System definition parameters, special cases, custom modules
 - [docs/generation.md](docs/generation.md) — `generate_key()` pipeline, CLI args, SVG→DXF, settings.scad composition
-- [docs/ui.md](docs/ui.md) — Tkinter UI screens, navigation, ProfileIndex, thumbnail caching, testing
+- [docs/data.md](docs/data.md) — ProfileIndex, profiles.json structure, testing
 - [docs/web.md](docs/web.md) — Flask web app pages, API endpoints, job store, template structure
 
 ## Commands
@@ -15,9 +15,6 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 All commands must be run in the venv.
 
 ```bash
-# Launch the Tkinter UI
-python autokey_ui.py
-
 # Launch the web app (http://localhost:5000)
 python web_app.py
 
@@ -30,23 +27,21 @@ python AutoKey.py --key 1,2,3,4,5 --profile profiles/AB-AB1.svg --definition def
 python -m pytest tests/
 
 # Run a single test class or test
-python -m pytest tests/test_app.py::TestProfileIndex
-python -m pytest tests/test_app.py::TestProfileDetail::test_combo_entry_enabled_in_key_mode
+python -m pytest tests/test_profile_index.py::TestProfileIndex
 
-# Install UI dependencies
-pip install -r requirements.txt   # Pillow only; svglib/reportlab are NOT needed
+# Install dependencies
+pip install -r requirements.txt
 ```
 
 ## Architecture
 
-### Three-layer design
+### Two-layer design
 
-The project has an original CLI layer, a Tkinter desktop UI, and a Flask web app:
+The project has a CLI layer and a Flask web app:
 
-- **`autokey_core.py`** — pure generation logic (extracted from the original `AutoKey.py`). Importable by both UIs. Key functions: `generate_key()`, `write_profile_scad()`, `write_system_scad()`.
+- **`autokey_core.py`** — pure generation logic (extracted from the original `AutoKey.py`). Key functions: `generate_key()`, `write_profile_scad()`, `write_system_scad()`.
 - **`AutoKey.py`** — thin CLI wrapper that delegates to `autokey_core.generate_key()`. Original interface preserved.
-- **`autokey_ui.py`** — entry point for the Tkinter desktop app.
-- **`ui/`** — all Tkinter UI code.
+- **`profile_index.py`** — data layer: loads/saves `profiles.json`, parses `.scad` files.
 - **`web_app.py`** — Flask web server (pages + REST API).
 - **`web/templates/`** — Jinja2 templates (extend `base.html`).
 - **`web/static/`** — static assets (Bootstrap loaded from CDN).
@@ -68,10 +63,6 @@ A key model requires **two independent files** used together:
 
 `settings.scad` is auto-generated on each run — do not edit it manually.
 
-### Tkinter UI navigation
-
-The app uses a single `Tk` window with a frame-stack. All screens (`HomeScreen`, `ProfileDetail`, `AddProfile`, `AddSystem`) are instantiated once at startup and raised via `frame.tkraise()`. `App.show_screen(name, **kwargs)` calls `frame.refresh(**kwargs)` before raising. The `App` instance holds the single `ProfileIndex` and is passed to every screen at construction.
-
 ### Web app navigation
 
 All pages extend `web/templates/base.html` (dark navbar with Bootstrap 5 + Bootstrap Icons). The `active_page` variable passed from each route highlights the current nav link.
@@ -90,18 +81,18 @@ Edit pages use a JS view/edit toggle (no page reload). Save calls `PUT /api/prof
 
 ### Data index
 
-`profiles.json` is the shared data source for both the Tkinter UI and the web app. It is auto-generated by `ProfileIndex.rebuild_from_disk()` on first run by scanning `.scad` files. Subsequent runs load it directly. It is updated incrementally when adding or editing profiles/systems.
+`profiles.json` is the data source for the web app. It is auto-generated by `ProfileIndex.rebuild_from_disk()` on first run by scanning `.scad` files. Subsequent runs load it directly. It is updated incrementally when adding or editing profiles/systems.
 
-`ProfileIndex` (`ui/profile_index.py`) key methods: `load()`, `save()`, `add_profile()`, `add_system()`, `update_profile()`, `update_system()`, `set_profile_default_system()`.
+`ProfileIndex` (`profile_index.py`) key methods: `load()`, `save()`, `add_profile()`, `add_system()`, `update_profile()`, `update_system()`, `set_profile_default_system()`.
 
 ### SVG thumbnails
 
-`ui/svg_renderer.py` uses Inkscape to convert SVGs to PNG thumbnails cached in `%TEMP%/autokey3d_thumbs/`. Pillow loads the PNGs into `ImageTk.PhotoImage`. No Cairo dependency. The web app serves the raw `.svg` files directly — browsers render SVG natively, so no conversion is needed server-side.
+The web app serves the raw `.svg` files directly — browsers render SVG natively, so no server-side conversion is needed.
 
 ### Testing
 
-`tests/test_app.py` uses a single module-level `tk.Tk()` root (Tkinter only supports one per process). UI screens are tested by constructing them with a `_MockApp` stub, driving them via `StringVar.set()` / direct method calls, and asserting on widget state. `autokey_core.generate_key` and `tkinter.messagebox` are patched to avoid real subprocess calls.
+`tests/test_profile_index.py` tests the `ProfileIndex` data layer. Tests use a temp directory with stub `.scad` files and cover: rebuild, field parsing, save/load roundtrip, add_profile, add_system.
 
 ## External tool dependencies
 
-OpenSCAD, Inkscape, and pstoedit must be on `PATH`. They are invoked as subprocesses. Inkscape is used for branding SVG→EPS→DXF conversion and for SVG thumbnail generation (Tkinter UI only). The pre-computed `.dxf` files in `profiles/` allow key generation without Inkscape for the profile step.
+OpenSCAD, Inkscape, and pstoedit must be on `PATH`. They are invoked as subprocesses. Inkscape is used for branding SVG→EPS→DXF conversion. The pre-computed `.dxf` files in `profiles/` allow key generation without Inkscape for the profile step.
