+++ profiles/AB-AB95.scad
@@ -1,10 +1,8 @@
 // The tolerance to use when removing material from the profile
-tol=0.2;
+tol = 0.2;
 
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
-ph=8.75;
-
-// Profile is thin, use a smaller key connector here to save some space
-khcx=2.5;
+ph=8.75 + 2*tol;
 profile_path = "profiles/AB-AB95.dxf";
+khcx=2.5;
