+++ profiles/AB-V14.scad
@@ -1,7 +1,7 @@
 // The tolerance to use when removing material from the profile
-tol=0.2;
+tol = 0.2;
 
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
-ph=9.30 + 2*tol;
+ph=9.3 + 2*tol;
 profile_path = "profiles/AB-V14.dxf";
