+++ profiles/BK-BK1.scad
@@ -3,5 +3,5 @@ tol = 0.2;
 
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
-ph=9 + 2*tol;
+ph=9.0 + 2*tol;
 profile_path = "profiles/BK-BK1.dxf";
