+++ profiles/BK-Serie-2000-Kaskade-Unbekannt.scad
@@ -3,5 +3,5 @@ tol = 0.2;
 
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
-ph = 8.6;
+ph=8.6 + 2*tol;
 profile_path = "profiles/BK-Serie-2000-Kaskade-Unbekannt.dxf";
