+++ profiles/CE-CE41.scad
@@ -1,5 +1,5 @@
 // The tolerance to use when removing material from the profile
-tol=0.2;
+tol = 0.2;
 
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
