+++ profiles/MD-CLASSIC.scad
@@ -4,8 +4,7 @@ tol = 0.25;
 // Key profile height (including tolerance, i.e. measured on the lock, not the blank)
 // If you have information on the key blank height, add 2*tol.
 ph=7.8 + 2*tol;
-
-khcz=7;
+profile_path = "profiles/MD-CLASSIC.dxf";
+khcz=7.0;
 thin_handle=true;
 match_handle=true;
-profile_path = "profiles/MD-CLASSIC.dxf";
